Initial Helper Directory created
Created helper directory that handles the heavy lifting of reading a
word doc and parsing out the necessary data that is used to create a
JSON file.

Refactor some of the Content classes.

Note: The current ToString() function does not work as expected. It
does not present the data that are contained in the variable correctly.